refactor: extract ProcessingRunOptions from run.rs, remove stale re-exports
Move ProcessingRunOptions to run/run_options.rs with pub(crate) visibility.
Remove #[allow(unused_imports)] re-export block that duplicated run_job
symbols and run_validation re-exports. Test module now imports directly
from super::run_job instead of transient re-exports.
